# Conflicts:
#	app/src/main/java/one/mixin/android/ui/home/MainActivity.kt
#	app/src/main/java/one/mixin/android/ui/home/web3/BrowserPage.kt
#	app/src/main/res/layout/fragment_deposit.xml
#	app/src/main/res/layout/fragment_transfer_balance_error_bottom_sheet.xml
#	app/src/main/res/layout/fragment_transfer_bottom_sheet.xml
#	app/src/main/res/layout/fragment_wallet_connect_bottom_sheet.xml
#	app/src/main/res/layout/fragment_web3_address.xml

+++ app/src/main/java/one/mixin/android/vo/safe/SafeSnapshot.kt
@@ -57,7 +57,7 @@ data class SafeSnapshot(
     @SerializedName("confirmations")
     @SerialName("confirmations")
     @ColumnInfo(name = "confirmations")
-    val confirmations: Int?,
+    val confirmations: Long?,
     @SerializedName("opening_balance")
     @SerialName("opening_balance")
     @ColumnInfo(name = "opening_balance")
